fix: moved copy curriculum
