+++ apps/web/src/app/(dashboard)/(management)/secretariaat/diplomalijn/cursussen/_components/action-buttons.tsx
@@ -13,7 +13,7 @@ import {
 import { Button } from "~/app/(dashboard)/_components/button";
 import { Input } from "~/app/(dashboard)/_components/input";
 import { useFormInput } from "~/app/_actions/hooks/useFormInput";
-import { copyCurriculumAction } from "~/app/_actions/secretariat/copy-curriculum-action";
+import { copyCurriculumAction } from "~/app/_actions/secretariat/curricula/copy-curriculum-action";
 import { DEFAULT_SERVER_ERROR_MESSAGE } from "~/app/_actions/utils";
 
 export function CopyCurriculum({ curriculumId }: { curriculumId: string }) {
